+++ StudentSoup/src/main/frontend/src/pages/signup/SignUpProcess2.tsx
@@ -1,7 +1,7 @@
 import React, { useCallback, useEffect, useState } from 'react';
 import { Desktop, Mobile } from 'mediaQuery';
 import Background from 'components/common/Background';
-import SignUpComponent from './SignUpComponent';
+import SignUpComponent from './components/SignUpComponent';
 import process_activate_2 from 'assets/images/signup_process_activate_2.png';
 import process_3 from 'assets/images/signup_process_3.png';
 import process_check from 'assets/images/signup_process_check.png';
