+++ StudentSoup/src/main/frontend/src/pages/signup/SignUpProcess3.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import Background from 'components/common/Background';
-import SignUpComponent from './SignUpComponent';
+import SignUpComponent from './components/SignUpComponent';
 import process_activate_3 from 'assets/images/signup_process_activate_3.png';
 import process_check from 'assets/images/signup_process_check.png';
 import process_next_bar from 'assets/images/signup_process_next_bar.png';
