+++ StudentSoup/src/main/frontend/src/pages/signup/components/Checkbox.tsx
