+++ StudentSoup/src/main/frontend/src/pages/signup/components/checkbox.scss
