+++ StudentSoup/src/main/frontend/src/pages/signup/components/signupcomponent.scss
